Some hosts may block file_put_contents like they did fwrite.

+++ Sources/Load.php
@@ -2723,17 +2723,18 @@ function cache_put_data($key, $value, $ttl = 120)
 			break;
 		default:
 			// Otherwise custom cache?
-			if ($value === null)
-				@unlink($cachedir . '/data_' . $key . '.php');
-			else
-			{
-				$cache_data = '<' . '?' . 'php if (!defined(\'SMF\')) die; if (' . (time() + $ttl) . ' < time()) $expired = true; else{$expired = false; $value = \'' . addcslashes($value, '\\\'') . '\';}' . '?' . '>';
-
-				// Write out the cache file, check that the cache write was successful; all the data must be written
-				// If it fails due to low diskspace, or other, remove the cache file
-				if (file_put_contents($cachedir . '/data_' . $key . '.php', $cache_data, LOCK_EX) !== strlen($cache_data))
+			if (function_exists('file_put_contents')) {
+				if ($value === null)
 					@unlink($cachedir . '/data_' . $key . '.php');
-			}
+				else
+					{
+					$cache_data = '<' . '?' . 'php if (!defined(\'SMF\')) die; if (' . (time() + $ttl) . ' < time()) $expired = true; else{$expired = false; $value = \'' . addcslashes($value, '\\\'') . '\';}' . '?' . '>';
+
+					// Write out the cache file, check that the cache write was successful; all the data must be written
+					// If it fails due to low diskspace, or other, remove the cache file
+					if (file_put_contents($cachedir . '/data_' . $key . '.php', $cache_data, LOCK_EX) !== strlen($cache_data))
+						@unlink($cachedir . '/data_' . $key . '.php');
+			}	}
 			break;
 	}
